# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AmendLastLogOut": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "log": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionLogOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "deleted"
        ],
        "title": "AmendLastLogOut",
        "type": "object"
      },
...
      "GoalOut": {
        "properties": {
          "is_streak_target": {
            "title": "Is Streak Target",
            "type": "boolean"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "target_value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Target Value",
            "type": "string"
          },
          "trackable_key": {
            "title": "Trackable Key",
            "type": "string"
          }
        },
        "required": [
          "trackable_key",
          "target_value",
          "period",
          "is_streak_target"
        ],
        "title": "GoalOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LogMealTemplateRequest": {
        "properties": {
          "logged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logged At"
          },
          "meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Type"
          },
          "multiplier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "1",
            "title": "Multiplier"
          }
        },
        "title": "LogMealTemplateRequest",
        "type": "object"
      },
      "LogNutritionRequest": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/NutritionEntryIn"
            },
            "minItems": 1,
            "title": "Entries",
            "type": "array"
          },
          "logged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logged At"
          },
          "meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "entries"
        ],
        "title": "LogNutritionRequest",
        "type": "object"
      },
      "LogOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "logged_at": {
            "format": "date-time",
            "title": "Logged At",
            "type": "string"
          },
          "meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "name",
          "logged_at",
          "meal_type"
        ],
        "title": "LogOut",
        "type": "object"
      },
...
      "MealTemplateSummaryOut": {
        "properties": {
          "default_meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Meal Type"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealTemplateItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "totals": {
            "additionalProperties": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "title": "Totals",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "default_meal_type",
          "items",
          "totals"
        ],
        "title": "MealTemplateSummaryOut",
        "type": "object"
      },
      "NutritionDayOut": {
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/TrackableProgressOut"
            },
            "title": "Bars",
            "type": "array"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "hero": {
            "$ref": "#/components/schemas/TrackableProgressOut"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/DayLogOut"
            },
            "title": "Logs",
            "type": "array"
          },
          "streak_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streak Key"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/MealTemplateSummaryOut"
            },
            "title": "Templates",
            "type": "array"
          }
        },
        "required": [
          "date",
          "hero",
          "bars",
          "streak_key",
          "logs",
          "templates"
        ],
        "title": "NutritionDayOut",
        "type": "object"
      },
...
      "NutritionLogOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "logged_at": {
            "format": "date-time",
            "title": "Logged At",
            "type": "string"
          },
          "meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "values": {
            "additionalProperties": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "logged_at",
          "meal_type",
          "values"
        ],
        "title": "NutritionLogOut",
        "type": "object"
      },
...
      "SaveMealTemplateRequest": {
        "properties": {
          "default_meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Meal Type"
          },
          "log_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Log Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "template_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "required": [
          "name",
          "log_ids"
        ],
        "title": "SaveMealTemplateRequest",
        "type": "object"
      },
      "SetGoalsRequest": {
        "properties": {
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalIn"
            },
            "minItems": 1,
            "title": "Goals",
            "type": "array"
          }
        },
        "required": [
          "goals"
        ],
        "title": "SetGoalsRequest",
        "type": "object"
      },
...
      "UpdateMealTemplateItemRequest": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "serving_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Description"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ]
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "name"
        ],
        "title": "UpdateMealTemplateItemRequest",
        "type": "object"
      },
      "UpdateNutritionLogRequest": {
        "properties": {
          "meal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "values": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                      "type": "string"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "title": "UpdateNutritionLogRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Swolemates",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/nutrition/day": {
      "get": {
        "operationId": "getNutritionDay",
        "parameters": [
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionDayOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Nutrition Day",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/goals": {
      "get": {
        "operationId": "getGoals",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "title": "Response Getgoals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Goals",
        "tags": [
          "nutrition"
        ]
      },
      "put": {
        "operationId": "setGoals",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetGoalsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "title": "Response Setgoals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Goals",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/logs": {
      "post": {
        "operationId": "logNutrition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogNutritionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Nutrition",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/logs/amend-last": {
      "post": {
        "operationId": "amendLastLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNutritionLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmendLastLogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Amend Last Log",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/logs/{log_id}": {
      "patch": {
        "operationId": "updateNutritionLog",
        "parameters": [
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Log Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNutritionLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionLogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Nutrition Log",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/templates": {
      "get": {
        "operationId": "listMealTemplates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MealTemplateSummaryOut"
                  },
                  "title": "Response Listmealtemplates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Meal Templates",
        "tags": [
          "nutrition"
        ]
      },
      "post": {
        "operationId": "saveMealTemplate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMealTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealTemplateSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Meal Template",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/templates/{template_id}": {
      "delete": {
        "operationId": "deleteMealTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Meal Template",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/templates/{template_id}/items/{item_id}": {
      "patch": {
        "operationId": "updateMealTemplateItem",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealTemplateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealTemplateSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Meal Template Item",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/api/nutrition/templates/{template_id}/log": {
      "post": {
        "operationId": "logMealTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogMealTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LogOut"
                  },
                  "title": "Response Logmealtemplate",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Meal Template",
        "tags": [
          "nutrition"
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "operationId": "getProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "profile"
        ]
      }
    },
...
  }
}